Rewrite SpectrumQueryBuilderAlt to search against both consensus and reference spectra

+++ src/main/java/org/dulab/adapcompounddb/site/repositories/SpectrumClusterRepositoryCustom.java
@@ -4,10 +4,11 @@ import org.dulab.adapcompounddb.models.entities.views.SpectrumClusterView;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigInteger;
 import java.util.Collection;
 
 public interface SpectrumClusterRepositoryCustom {
 
     Page<SpectrumClusterView> findClusters(
-            String searchStr, Iterable<Long> submissionIds, Pageable pageable);
+            String searchStr, Iterable<BigInteger> submissionIds, Pageable pageable);
 }
